```
benchmark                               old ns/op       new ns/op     delta
Benchmark_T1000_P5000_Matches-4         2614306617      10704746      -99.59%
Benchmark_T1000_P5000_NoMatches-4       2515589915      10024400      -99.60%
Benchmark_T100_P5000_Matches-4          33235249        13528192      -59.30%
Benchmark_T100_P5000_NoMatches-4        35301480        2377025       -93.27%
Benchmark_T10_P5000_Matches-4           3999232         2545317       -36.35%
Benchmark_T10_P5000_NoMatches-4         4069072         2356547       -42.09%
Benchmark_T10_P500_CountTask-4          4308417         556166        -87.09%
Benchmark_T10_P500000_CountTask-4       171780120       154546757     -10.03%
Benchmark_T1000_P500-4                  17486595152     15440082      -99.91%
Benchmark_T1000_P500000_CountTask-4     19294813464     318064303     -98.35%

benchmark                               old allocs     new allocs     delta
Benchmark_T1000_P5000_Matches-4         12084          10060          -16.75%
Benchmark_T1000_P5000_NoMatches-4       12083          10061          -16.73%
Benchmark_T100_P5000_Matches-4          1264           1073           -15.11%
Benchmark_T100_P5000_NoMatches-4        1266           1055           -16.67%
Benchmark_T10_P5000_Matches-4           179            155            -13.41%
Benchmark_T10_P5000_NoMatches-4         178            155            -12.92%
Benchmark_T10_P500_CountTask-4          388            336            -13.40%
Benchmark_T10_P500000_CountTask-4       428            385            -10.05%
Benchmark_T1000_P500-4                  32184          27166          -15.59%
Benchmark_T1000_P500000_CountTask-4     32253          27199          -15.67%

benchmark                               old bytes     new bytes     delta
Benchmark_T1000_P5000_Matches-4         1818384       1752617       -3.62%
Benchmark_T1000_P5000_NoMatches-4       1818384       1752675       -3.61%
Benchmark_T100_P5000_Matches-4          1413715       1408155       -0.39%
Benchmark_T100_P5000_NoMatches-4        1413848       1406806       -0.50%
Benchmark_T10_P5000_Matches-4           1373087       1372209       -0.06%
Benchmark_T10_P5000_NoMatches-4         1372988       1372138       -0.06%
Benchmark_T10_P500_CountTask-4          157302        155606        -1.08%
Benchmark_T10_P500000_CountTask-4       149610906     149609495     -0.00%
Benchmark_T1000_P500-4                  1433632       1272199       -11.26%
Benchmark_T1000_P500000_CountTask-4     150886920     150723879     -0.11%
```
_Tested against master, with edgeBufferSize of 1000 (default value)_