@yosiat 2016-04-01
@nathanielc As written in #360 -

> I think there is something strange with the updateKeys/benchmark, it's only callers are NewStatistics and DeleteStatistics, and the benchmark shouldn't time them.
> 
> But I will submit PR to remove it, the only usage of the keys is in doLocked:
> ```go
> func (v *Map) doLocked(f func(expvar.KeyValue)) {
>     for _, k := range v.keys {
>         f(expvar.KeyValue{k, v.m[k]})
>     }
> }
> ```
> I don't understand why not to use "m" already?

I didn't find any good reason for not using v.m for iterating the map in doLocked, so I switched to iterating the map directly.

I still don't understand why this change caused a dramatic change in this benchmark, since updateKeys shouldn't be called unless it's NewStatistics/DeleteStatistics