@yosiat 2016-04-01
It's breaking becuase:
```
3310 tick/cmd/tickdoc/main.go:133: not enough arguments in call to markdown.NewRenderer^M
```

@nathanielc  ?