@yosiat 2016-04-01
And it looks that updateKeys is called a lot in this profile because of Drain - whom calls DeleteStatistics eventually.